# docs/openapi.yaml
openapi: 3.0.3
info:
  title: India In-Time API
  version: "3.6.0"
  description: |
    Production API contract for India In-Time. Request and response schemas are
    intentionally explicit so frontend, mobile, QA, and third-party clients can
    validate payloads without reverse-engineering route implementations.
servers:
  - url: /api
    description: Deployment-relative API root
...
paths:
...
  /weather:
    get:
      tags: [Weather]
      summary: Weather proxy
      parameters:
        - { name: lat, in: query, required: true, schema: { type: number, minimum: -90, maximum: 90 } }
        - { name: lon, in: query, required: true, schema: { type: number, minimum: -180, maximum: 180 } }
      responses:
        '200':
          description: Weather payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WeatherResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /weather/alerts:
    get:
      tags: [Weather]
      summary: Weather and severe-condition alerts
      parameters:
        - { name: city, in: query, required: false, schema: { type: string, maxLength: 100 } }
        - { name: lat, in: query, required: false, schema: { type: number, minimum: -90, maximum: 90 } }
        - { name: lon, in: query, required: false, schema: { type: number, minimum: -180, maximum: 180 } }
      responses:
        '200':
          description: Alert payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WeatherAlertsResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /trips:
    get:
      tags: [Trips]
      summary: List current user's saved trips
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        '200':
          description: Saved trips
          content:
            application/json:
              schema:
                type: object
                properties:
                  trips:
                    type: array
                    items: { $ref: '#/components/schemas/Trip' }
                required: [trips]
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Trips]
      summary: Save a trip
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TripCreate' }
      responses:
        '201':
          description: Trip created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Trip' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /trips/{id}:
    parameters:
      - { $ref: '#/components/parameters/TripId' }
    get:
      tags: [Trips]
      summary: Get one saved trip owned by the caller
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Trip
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Trip' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Trips]
      summary: Delete one saved trip owned by the caller
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  parameters:
...
    TripId:
      name: id
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        requestId: { type: string }
...
    Coordinates:
      type: array
      minItems: 2
      maxItems: 2
      items: { type: number }
      example: [17.6868, 83.2185]
...
    WeatherResponse:
      type: object
      properties:
        temperature: { type: number }
        feelsLike: { type: number }
        humidity: { type: number, minimum: 0, maximum: 100 }
        weather: { type: string }
        description: { type: string }
        windSpeed: { type: number }
        observedAt: { type: string, format: date-time }
    WeatherAlertsResponse:
      type: object
      required: [alerts]
      properties:
        alerts:
          type: array
          items:
            type: object
            properties:
              severity: { type: string, enum: [info, watch, warning, emergency] }
              title: { type: string }
              message: { type: string }
              startsAt: { type: string, format: date-time }
...
    TripCreate:
      type: object
      required: [city, plan]
      properties:
        city: { type: string, minLength: 1, maxLength: 100 }
        startTime: { type: string, pattern: '^\\d{2}:\\d{2}$' }
        durationMinutes: { type: integer, minimum: 1, maximum: 1440 }
        plan:
          type: array
          maxItems: 30
          items:
            type: array
            items: { $ref: '#/components/schemas/TripStop' }
        budget: { type: number, minimum: 0 }
        metadata: { type: object, additionalProperties: true }
    Trip:
      allOf:
        - $ref: '#/components/schemas/TripCreate'
        - type: object
          required: [id]
          properties:
            id: { type: string }
            userId: { type: string }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    TripStop:
      type: object
      required: [name]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        coords: { $ref: '#/components/schemas/Coordinates' }
        startTime: { type: string }
        endTime: { type: string }
        visitMinutes: { type: integer, minimum: 0 }
...